requirements.txt
# Computer Vision and AI
opencv-python>=4.8.0
torch>=1.13.0
torchvision>=0.14.0
yolov5>=7.0.0

# Web Framework
Flask>=2.3.0

# System and Hardware
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"
psutil>=5.9.0

# Communication
requests>=2.31.0

# Data Processing
numpy>=1.24.0

# Configuration
python-dotenv>=1.0.0

# Optional: Motion detection (if using VIBE background subtraction)
# vibe>=0.1.0  # Uncomment if using motion detection